``` feat(gui): 为 select_menu 添加默认选中项支持
- 在 `select_menu` 函数中新增 `default_selected` 参数，用于指定默认选中的菜单项
- 自动调整初始页面位置以确保默认选中项可见
- 修复按钮长按检测逻辑，使用按钮对象的 pin 方法替代宏定义

feat(serial): 增加带颜色的串口打印宏定义

- 新增多种颜色打印宏（ERROR_COLOR、WARNING_COLOR 等）
- 提供便捷的打印封装宏（PRINT_ERROR、PRINT_WARNING 等）

feat(hal): 优化按钮钩子函数，支持非等待模式

- 在 `hookButton` 和 `unhookButton` 中增加 `no_wait` 参数
- 可跳过按钮状态等待过程，提升响应速度

feat(peripherals): 添加 BQ27441 电池管理芯片支持

- 定义 PERIPHERALS_BQ27441_BIT 用于外设检测

feat(bq27441): 调整 I2C 超时时间

- 将 BQ72441_I2C_TIMEOUT 从 2000ms 缩短至 250ms

feat(ds3231): 使用 uint64_t 类型提升 unixtime 精度

- 修改 `unixtime()` 返回值类型为 `uint64_t`，防止时间溢出

feat(audio): 增强 FLAC 和 Vorbis 格式元数据解析能力(不可用状态)

- 新增 AudioFileSourceVorbis 类支持 FLAC/Opus 元数据读取
- 实现 Vorbis 注释解析和映射到标准标签
- 修复 AudioGeneratorFLAC 中被注释掉的元数据处理逻辑

feat(i2s): 修正 I2S 音频输出引脚和参数设置

- 更改 doutPin 引脚编号为 21
- 修复 Set_bits_per_chan 参数命名冲突问题

feat(web): 增加对 FLAC 文件的 MIME 类型支持

- 在 AsyncFileResponse 中添加 `.flac` 文件类型识别

feat(epd): 增强 GxEPD2 屏幕驱动兼容性(目前仍然有问题)

- 支持 T5D 模式切换
- 增加自定义波形查找表 (LUT)
- 调整刷新延迟参数以适配不同模式
```

+++ lib/DS3231/DS3231.cpp
@@ -142,8 +142,8 @@ static uint8_t conv2d(const char* p) {
     return 10 * v + *++p - '0';
 }
 // UNIX time: IS CORRECT ONLY WHEN SET TO UTC!!!
-uint32_t DateTime::unixtime(void) const {
-  uint32_t t;
+uint64_t DateTime::unixtime(void) const {
+  uint64_t t;
   uint16_t days = date2days(yOff, m, d);
   t = time2long(days, hh, mm, ss);
   t += SECONDS_FROM_1970_TO_2000;  // seconds from 1970 to 2000
